\name{Classify.dropout-package}
\alias{Classify.dropout-package}
\docType{package}
\title{
\packageTitle{Classify.dropout}
}
\description{
\packageDescription{Classify.dropout}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Classify.dropout}
\packageIndices{Classify.dropout}
}
\author{
\packageAuthor{Classify.dropout}

Maintainer: \packageMaintainer{Classify.dropout}
}
\references{

}

\keyword{ package }
\seealso{
}
\examples{
data(db_example)
classify.dropout(db_example)
cart_fit$finalModel
mc_cart
}
