\name{Classify.dropout-package}
\alias{Classify.dropout-package}
\alias{Classify.dropout}
\docType{package}
\title{
\packageTitle{Classify.dropout}
}
\description{
\packageDescription{Classify.dropout}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Classify.dropout}
\packageIndices{Classify.dropout}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Classify.dropout}

Maintainer: \packageMaintainer{Classify.dropout}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
